BlumBot splitted into separate files
