+++ Blum/Models/BlumAccessTokenJson.cs
@@ -1,0 +1,18 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Blum.Models
+{
+    public class BlumAccessTokenJson
+    {
+        public class TokenInfo
+        {
+            [JsonPropertyName("access")]
+            public string? Access { get; set; }
+
+            [JsonPropertyName("refresh")]
+            public string? Refresh { get; set; }
+        }
+        [JsonPropertyName("token")]
+        public TokenInfo? Token { get; set; }
+    }
+}
